src/MyFunction/requirements.txt
python-memcached==1.59

